Admin server refactors
