+++ products/jbrowse-cli/bin/run
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../dist/index.js').main(process.argv.slice(2))
+require('../lib/index.js').main(process.argv.slice(2))
