+++ products/jbrowse-cli/src/commands/create.test.ts
@@ -7,7 +7,7 @@ import path from 'path'
 
 import nock from 'nock'
 
-import { runInTmpDir, runNativeCommand } from '../testUtil'
+import { runInTmpDir, runCommand } from '../testUtil'
 
 const { readdir } = fs.promises
 
@@ -40,20 +40,20 @@ afterAll(() => (process.exitCode = 0))
 
 test('fails if no path is provided to the command', async () => {
   await runInTmpDir(async () => {
-    const { error } = await runNativeCommand(['create'])
+    const { error } = await runCommand(['create'])
     expect(error?.message).toMatchSnapshot()
   })
 })
 
 test('fails if no path is provided to the command, even with force', async () => {
   await runInTmpDir(async () => {
-    const { error } = await runNativeCommand(['create', '--force'])
+    const { error } = await runCommand(['create', '--force'])
     expect(error?.message).toMatchSnapshot()
   })
 })
 
 test('fails if user selects a directory that has existing files', async () => {
-  const { error } = await runNativeCommand(['create', '.'])
+  const { error } = await runCommand(['create', '.'])
   expect(error?.message).toMatchSnapshot()
 })
 test('fails if the fetch does not return the right file', async () => {
@@ -61,7 +61,7 @@ test('fails if the fetch does not return the right file', async () => {
     nock('https://example.com')
       .get('/jbrowse-web-v0.0.1.zip')
       .reply(200, 'I am the wrong type', { 'Content-Type': 'application/json' })
-    const { error } = await runNativeCommand([
+    const { error } = await runCommand([
       'create',
       'jbrowse',
       '--url',
@@ -77,7 +77,7 @@ test('download and unzips to new directory', async () => {
       .get('/repos/GMOD/jbrowse-components/releases?page=1')
       .reply(200, releaseArray)
     mockZip()
-    await runNativeCommand(['create', 'jbrowse'])
+    await runCommand(['create', 'jbrowse'])
     expect(await readdir(path.join(ctx.dir, 'jbrowse'))).toContain(
       'manifest.json',
     )
@@ -87,7 +87,7 @@ test('download and unzips to new directory', async () => {
 test('downloads from a url', async () => {
   await runInTmpDir(async ctx => {
     mockZip()
-    await runNativeCommand([
+    await runCommand([
       'create',
       'jbrowse',
       '--url',
@@ -105,7 +105,7 @@ test('overwrites and succeeds in download in a non-empty directory with tag', as
       .get('/repos/GMOD/jbrowse-components/releases/tags/v0.0.1')
       .reply(200, releaseArray[0])
     mockZip()
-    await runNativeCommand(['create', 'jbrowse', '--tag', 'v0.0.1', '--force'])
+    await runCommand(['create', 'jbrowse', '--tag', 'v0.0.1', '--force'])
     expect(await readdir(path.join(ctx.dir, 'jbrowse'))).toContain(
       'manifest.json',
     )
@@ -117,7 +117,7 @@ test('fails to download a version that does not exist', async () => {
     nock('https://api.github.com')
       .get('/repos/GMOD/jbrowse-components/releases/tags/v999.999.999')
       .reply(404, {})
-    const { error } = await runNativeCommand([
+    const { error } = await runCommand([
       'create',
       'jbrowse',
       '--tag',
@@ -134,8 +134,8 @@ test('fails because this directory is already set up', async () => {
       .get('/repos/GMOD/jbrowse-components/releases?page=1')
       .reply(200, releaseArray)
     mockZip()
-    await runNativeCommand(['create', 'jbrowse'])
-    const { error } = await runNativeCommand(['create', 'jbrowse'])
+    await runCommand(['create', 'jbrowse'])
+    const { error } = await runCommand(['create', 'jbrowse'])
     expect(error?.message).toMatchSnapshot()
   })
 })
@@ -147,7 +147,7 @@ test('lists versions', async () => {
       .reply(200, releaseArray)
       .get('/repos/GMOD/jbrowse-components/releases?page=2')
       .reply(200, [])
-    const { stdout } = await runNativeCommand(['create', '--listVersions'])
+    const { stdout } = await runCommand(['create', '--listVersions'])
     expect(stdout).toMatchSnapshot()
   })
 })
