+++ products/jbrowse-cli/src/commands/make-pif.test.ts
@@ -6,7 +6,7 @@ import fs from 'fs'
 import path from 'path'
 import { gunzipSync } from 'zlib'
 
-import { runInTmpDir, runNativeCommand } from '../testUtil'
+import { runInTmpDir, runCommand } from '../testUtil'
 
 const base = path.join(__dirname, '..', '..', 'test', 'data')
 const simplePaf = path.join(base, 'volvox_inv_indels.paf')
@@ -16,7 +16,7 @@ const exists = (p: string) => fs.existsSync(p)
 test('make-pif', async () => {
   await runInTmpDir(async () => {
     const fn = `${path.basename(simplePaf, '.paf')}.pif.gz`
-    await runNativeCommand(['make-pif', simplePaf, '--out', fn])
+    await runCommand(['make-pif', simplePaf, '--out', fn])
     expect(exists(fn)).toBeTruthy()
     expect(gunzipSync(fs.readFileSync(fn)).toString()).toMatchSnapshot()
   })
@@ -25,7 +25,7 @@ test('make-pif', async () => {
 test('make pif with CSI', async () => {
   await runInTmpDir(async () => {
     const fn = `${path.basename(simplePaf, '.paf')}.pif.gz`
-    await runNativeCommand(['make-pif', simplePaf, '--out', fn, '--csi'])
+    await runCommand(['make-pif', simplePaf, '--out', fn, '--csi'])
     expect(exists(fn)).toBeTruthy()
     expect(exists(`${fn}.csi`)).toBeTruthy()
   })
