+++ products/jbrowse-cli/src/commands/remove-track.test.ts
@@ -5,7 +5,7 @@
 import fs from 'fs'
 import path from 'path'
 
-import { readConf, runInTmpDir, runNativeCommand } from '../testUtil'
+import { readConf, runInTmpDir, runCommand } from '../testUtil'
 
 const twoPath = path.join(__dirname, '..', '..', 'test', 'data', 'simple.2bit')
 
@@ -16,9 +16,9 @@ afterAll(() => (process.exitCode = 0))
 test('remove track', async () => {
   await runInTmpDir(async ctx => {
     fs.copyFileSync(twoPath, path.join(ctx.dir, path.basename(twoPath)))
-    await runNativeCommand(['add-assembly', 'simple.2bit', '--load', 'copy'])
-    await runNativeCommand(['add-track', 'simple.bam', '--load', 'inPlace'])
-    await runNativeCommand(['remove-track', 'simple'])
+    await runCommand(['add-assembly', 'simple.2bit', '--load', 'copy'])
+    await runCommand(['add-track', 'simple.bam', '--load', 'inPlace'])
+    await runCommand(['remove-track', 'simple'])
     const contents = readConf(ctx)
     expect(contents.assemblies.length).toBe(1)
     expect(contents.tracks.length).toBe(0)
